move connection execution to `execution.rs`
